+++ selene-lib/src/rules/unused_variable.rs
@@ -1,3 +1,8 @@
+use crate::{
+    ast_util::scopes::AssignedValue,
+    standard_library::{Field, FieldKind, Observes},
+};
+
 use super::*;
 
 use full_moon::ast::Ast;
@@ -25,6 +30,13 @@ pub struct UnusedVariableLint {
     ignore_pattern: Regex,
 }
 
+#[derive(Debug, PartialEq)]
+pub enum AnalyzedReference {
+    Read,
+    PlainWrite,
+    ObservedWrite(Label),
+}
+
 impl Rule for UnusedVariableLint {
     type Config = UnusedVariableConfig;
     type Error = regex::Error;
@@ -36,7 +48,7 @@ impl Rule for UnusedVariableLint {
         })
     }
 
-    fn pass(&self, _: &Ast, _: &Context, ast_context: &AstContext) -> Vec<Diagnostic> {
+    fn pass(&self, _: &Ast, context: &Context, ast_context: &AstContext) -> Vec<Diagnostic> {
         let mut diagnostics = Vec::new();
 
         for (_, variable) in ast_context
@@ -45,13 +57,80 @@ impl Rule for UnusedVariableLint {
             .iter()
             .filter(|(_, variable)| !self.ignore_pattern.is_match(&variable.name))
         {
-            let mut references = variable
+            let references = variable
                 .references
                 .iter()
                 .copied()
                 .map(|id| &ast_context.scope_manager.references[id]);
 
-            if !references.clone().any(|reference| reference.read) {
+            // We need to make sure that references that are marked as "read" aren't only being read in an "observes: write" context.
+            let analyzed_references = references
+                .map(|reference| {
+                    if reference.write {
+                        return AnalyzedReference::PlainWrite;
+                    }
+
+                    if variable.value != Some(AssignedValue::StaticTable) {
+                        return AnalyzedReference::Read;
+                    }
+
+                    let within_function_stmt = match &reference.within_function_stmt {
+                        Some(within_function_stmt) => within_function_stmt,
+                        None => return AnalyzedReference::Read,
+                    };
+
+                    let function_call_stmt = &ast_context.scope_manager.function_calls
+                        [within_function_stmt.function_call_stmt_id];
+
+                    // The function call it's within is script defined, we can't assume anything
+                    if ast_context.scope_manager.references[function_call_stmt.initial_reference]
+                        .resolved
+                        .is_some()
+                    {
+                        return AnalyzedReference::Read;
+                    }
+
+                    let function_behavior = match context
+                        .standard_library
+                        .find_global(&function_call_stmt.call_name_path)
+                    {
+                        Some(Field {
+                            field_kind: FieldKind::Function(function_behavior),
+                            ..
+                        }) => function_behavior,
+                        _ => return AnalyzedReference::Read,
+                    };
+
+                    let argument = match function_behavior
+                        .arguments
+                        .get(within_function_stmt.argument_index)
+                    {
+                        Some(argument) => argument,
+                        None => return AnalyzedReference::Read,
+                    };
+
+                    let write_only = argument.observes == Observes::Write;
+
+                    if !write_only {
+                        return AnalyzedReference::Read;
+                    }
+
+                    AnalyzedReference::ObservedWrite(Label::new_with_message(
+                        reference.identifier,
+                        format!(
+                            "`{}` only writes to `{}`",
+                            // TODO: This is a typo if this is a method call
+                            function_call_stmt.call_name_path.join("."),
+                            variable.name
+                        ),
+                    ))
+                })
+                .collect::<Vec<_>>();
+
+            if !analyzed_references
+                .iter()
+                .any(|reference| reference == &AnalyzedReference::Read)
+            {
                 let mut notes = Vec::new();
 
                 if variable.is_self {
@@ -64,16 +143,27 @@ impl Rule for UnusedVariableLint {
                         .push("if you don't need it, consider using `.` instead of `:`".to_owned());
                 }
 
+                let write_only = !analyzed_references.is_empty();
+
                 diagnostics.push(Diagnostic::new_complete(
                     "unused_variable",
-                    if references.any(|reference| reference.write) {
+                    if write_only {
                         format!("{} is assigned a value, but never used", variable.name)
                     } else {
                         format!("{} is defined, but never used", variable.name)
                     },
                     Label::new(variable.identifiers[0]),
                     notes,
-                    Vec::new(),
+                    analyzed_references
+                        .into_iter()
+                        .filter_map(|reference| {
+                            if let AnalyzedReference::ObservedWrite(label) = reference {
+                                Some(label)
+                            } else {
+                                None
+                            }
+                        })
+                        .collect(),
                 ));
             };
         }
@@ -181,6 +271,15 @@ mod tests {
             UnusedVariableLint::new(UnusedVariableConfig::default()).unwrap(),
             "unused_variable",
             "objects",
+        );
+    }
+
+    #[test]
+    fn test_observes() {
+        test_lint(
+            UnusedVariableLint::new(UnusedVariableConfig::default()).unwrap(),
+            "unused_variable",
+            "observes",
         );
     }